+++ retail/BetterBlizzPlates.lua
@@ -374,6 +374,8 @@ local defaultSettings = {
     questIndicatorYPos = 0,
     questIndicatorAnchor = "LEFT",
     questIndicatorTestMode = false,
+    questIndicatorColorNpc = false,
+    questIndicatorColorNpcRGB = {1, 0.502, 0, 1},
     -- Font and texture
     customFontSizeEnabled = false,
     customFontSize = 12,
@@ -3504,6 +3506,10 @@ hooksecurefunc("CompactUnitFrame_UpdateHealthColor", function(frame)
         ColorNameplateByReaction(frame)
     end
 
+    if frame.isQuestNpc then
+        frame.healthBar:SetStatusBarColor(unpack(frame.isQuestNpc))
+    end
+
     if config.colorNPC then--and config.npcHealthbarColor then --bodify need npc check here since it can run before np added
         --frame.healthBar:SetStatusBarColor(config.npcHealthbarColor.r, config.npcHealthbarColor.g, config.npcHealthbarColor.b)
         BBP.ColorNpcHealthbar(frame)
@@ -3668,6 +3674,10 @@ function BBP.CompactUnitFrame_UpdateHealthColor(frame, exitLoop)
         if not exitLoop then
             ColorNameplateByReaction(frame)
         end
+    end
+
+    if frame.isQuestNpc then
+        frame.healthBar:SetStatusBarColor(unpack(frame.isQuestNpc))
     end
 
     if config.colorNPC and config.npcHealthbarColor then
@@ -4448,6 +4458,7 @@ local function HandleNamePlateRemoved(unit)
     if not frame then return end
 
     frame.bbpHiddenNPC = nil
+    frame.isQuestNpc = nil
     frame.ciChange = nil
     frame:SetScale(1)
     frame:SetAlpha(1)